+++ open_instruct/ppo_vllm_thread_ray_gtrl.py
@@ -122,7 +122,7 @@ from open_instruct.utils import (
     maybe_use_ai2_wandb_entity,
     setup_logger,
 )
-from open_instruct.vllm_utils3 import create_vllm_engines, init_process_group
+from open_instruct.vllm_utils import create_vllm_engines, init_process_group
 
 api = HfApi()
 INVALID_LOGPROB = 1.0
